server/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#

django-cors-headers==1.1.0
Django==1.9.1
djangorestframework-jwt==1.8.0
djangorestframework==3.3.2
psycopg2==2.6.1
Pillow==3.2.0
PyJWT==1.4.0              # via djangorestframework-jwt
